move ProtocolList and SessionList to Setup
