+++ src/containers/Setup/index.js
@@ -1,3 +1,5 @@
 export { default as ProtocolImport } from './ProtocolImport';
 export { default as ServerList } from './ServerList';
 export { default as SetupScreen } from './SetupScreen';
+export { default as ProtocolList } from './ProtocolList';
+export { default as SessionList } from './SessionList';
