+++ src/containers/index.js
@@ -15,5 +15,3 @@ export { default as ConcentricCircles } from './ConcentricCircles';
 export { default as NodeBucket } from './NodeBucket';
 export { default as LoadScreen } from './LoadScreen';
 export { default as LoadParamsRoute } from './LoadParamsRoute';
-export { default as ProtocolList } from './ProtocolList';
-export { default as SessionList } from './SessionList';
